Allow offline players to be removed from hunters list; update plugin.yml; rename packages
Instead of storing the hunters as a set of their usernames, we use a set of their UUIDs. Since players can change their usernames, but not their UUIDs, a set of UUIDs is better than a set of usernames for storage. When removing a hunter using /removehunter <username>, the given username is converted to a UUID using the Mojang API, then the resulting UUID is removed from the set of hunter UUIDs.

The packages have been renamed because if two plugins happen to have the same path to their respective main classes, only one of those plugins will be initialized at runtime.